requirements.txt
requests
beautifulsoup4
fastapi
uvicorn
python-dateutil